Categoria Hecho. Cesta: Falta controller

+++ backend/src/main/java/com/miguel/vendix/business/model/Producto.java
@@ -22,7 +22,7 @@ public class Producto {
 	
 	private String nombre;
 	private String descripcion;
-	private float precio;
+	private Double precio;
 	private int stock;
 	private boolean disponible;
